requirements-combined.txt
# B9 Dashboard - Combined Requirements for API + Scraper
# Production deployment on Render with both services

# =============================================================================
# SHARED DEPENDENCIES (Used by both API and Scraper)
# =============================================================================

# Core Async Support
asyncio-throttle>=1.0.2,<2.0.0
anyio>=3.7.0,<4.0.0
aiohttp>=3.9.1,<4.0.0
aiofiles>=23.2.1,<24.0.0

# Reddit API (Both use this)
asyncpraw>=7.7.1,<8.0.0
asyncprawcore>=2.4.0,<3.0.0

# Database - Supabase
postgrest==0.13.2
gotrue==2.5.4
supabase==2.3.4
realtime==1.0.0
storage3==0.6.0

# Redis for caching and queuing
redis>=5.0.1,<6.0.0
hiredis>=2.2.3,<3.0.0

# Environment & Configuration
python-dotenv>=1.0.0,<2.0.0

# JSON Processing
orjson>=3.9.10,<4.0.0

# Date/Time
python-dateutil>=2.8.2,<3.0.0

# User Agent Generation
fake-useragent>=1.4.0,<2.0.0

# Logging
structlog>=24.1.0,<25.0.0
loguru>=0.7.0,<1.0.0

# Type hints
typing-extensions>=4.9.0,<5.0.0

# =============================================================================
# API SPECIFIC (FastAPI Web Service)
# =============================================================================

# Web Framework
fastapi>=0.104.1,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
python-multipart>=0.0.6
starlette>=0.27.0,<1.0.0

# HTTP Client
httpx>=0.25.2,<1.0.0
requests>=2.31.0,<3.0.0

# OpenAI Integration
openai>=1.3.0,<2.0.0

# Rate Limiting & Security
slowapi>=0.1.9,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
cryptography>=41.0.0,<42.0.0

# Pydantic for validation
pydantic>=2.4.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# Performance Monitoring
psutil>=5.9.0,<6.0.0
prometheus-client>=0.19.0,<1.0.0

# Background Jobs
rq>=1.15.0,<2.0.0
schedule>=1.2.0,<2.0.0

# Database drivers
asyncpg>=0.29.0,<1.0.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0

# Health Monitoring
healthcheck>=1.3.3,<2.0.0

# Compression
brotli>=1.1.0,<2.0.0
zstandard>=0.21.0,<1.0.0

# Error Tracking
sentry-sdk[fastapi]>=1.38.0,<2.0.0

# =============================================================================
# SCRAPER SPECIFIC (Reddit Scraper)
# =============================================================================

# Retry and resilience
tenacity>=8.2.3,<9.0.0
circuitbreaker>=2.0.0,<3.0.0

# Rate limiting
ratelimit>=2.2.1,<3.0.0

# Proxy support
socksio>=1.0.0,<2.0.0

# =============================================================================
# PROCESS MANAGEMENT (For running both services)
# =============================================================================

# Supervisor for managing multiple processes
supervisor>=4.2.5,<5.0.0

# =============================================================================
# DEVELOPMENT ONLY (Optional - not installed in production)
# =============================================================================

# Testing
pytest>=7.4.3,<8.0.0; python_version>="3.8" and sys_platform!="win32"
pytest-asyncio>=0.21.1,<1.0.0; python_version>="3.8" and sys_platform!="win32"
pytest-mock>=3.12.0,<4.0.0; python_version>="3.8" and sys_platform!="win32"

# Code formatting
black>=23.7.0,<24.0.0; python_version>="3.8" and sys_platform!="win32"
isort>=5.12.0,<6.0.0; python_version>="3.8" and sys_platform!="win32"

# Type checking
types-redis>=4.6.0
types-requests>=2.31.0
types-python-dateutil>=2.8.0